Wifi-1502 : Hostapd has the list of channels it can switch to in case of DFS event
- Hostapd will have a list of channels which it is allowed to switch to after receiving DFS event.
- cpp check switched from 1.82 to 1.90-4 so some code changes in redirclient.c and captive.c for passing static analysis.

+++ feeds/wlan-ap/opensync/src/platform/openwrt/src/command/src/redirclient.c
@@ -232,7 +232,7 @@ static int redirLinkClientPongReceiveHandler(struct lws *wsi, struct per_session
 	l = 0;
 
 	while (shift >= 0) {
-		l |= (*p++) << shift;
+		l |= ((unsigned long)(*p++)) << shift;
 		shift -= 8;
 	}
